move timing functions into a general location
