+++ src/globals/index.ts
@@ -125,3 +125,6 @@ export const LOW_GAS_ERROR = 'TRANSACTION CANCELLED: Gas limit set too low. Plea
 export const DEFAULT_ERROR = 'TRANSACTION CANCELLED: Please check your developer console for more information.'
 
 export const WATCHER_INTERVAL = 5000
+
+export const AUCTION_RUN_TIME = 6.5 * 60 * 60 * 1000 // 6.5 hours in ms
+export const WAITING_PERIOD = 10 * 60 * 1000 // 10 min in ms
